Modify result page
If the user is logged in, landing on the result page triggers a gql
mutation request (createOrder). This request takes data from the shopping
cart and submits it to the server, which saves the data in the database.
Not using SSR prevents the CheckoutResult component from being rendered
twice, with each render submitting its own gql mutation request.